refactor: 3단계 미션 코드 정리
- Car의 리스트를 갖는 1급 컬렉션 Cars 작성
- domain, view 패키지 분리
