+++ src/main/java/car_racing/view/CarPrinter.java
@@ -1,13 +1,14 @@
-package car_racing;
+package car_racing.view;
 
-import java.util.List;
+import car_racing.domain.model.Car;
+import car_racing.domain.model.Cars;
 
 public class CarPrinter {
     public static void showRaceStart() {
         System.out.println("\n실행 결과");
     }
 
-    public static void showRaceResult(List<Car> cars) {
+    public static void showRaceResult(Cars cars) {
         cars.forEach(CarPrinter::showEachRaceResult);
         System.out.println();
     }
